# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Document API",
        "title": "Geolocation",
        "contact": {
            "name": "API Support",
            "url": "https://simpplify.com.br/contact",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/driver/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cria um motorista.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Motoristas"
                ],
                "summary": "Criar um Motorista.",
                "parameters": [
                    {
                        "description": "Requisição de Motorista",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/drivers.CreateDriverRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações do Motorista",
                        "schema": {
                            "$ref": "#/definitions/drivers.DriverResponse"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/driver/delete/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Exclui um motorista.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Motoristas"
                ],
                "summary": "Excluir um Motorista.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Motorista",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/driver/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Recupera as informações do motorista.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Motoristas"
                ],
                "summary": "Obter Motorista.",
                "responses": {
                    "200": {
                        "description": "Informações do Motorista",
                        "schema": {
                            "$ref": "#/definitions/drivers.DriverResponse"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/driver/list/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Recupera as informações do motorista a partir do ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Motoristas"
                ],
                "summary": "Obter Motorista por ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Motorista",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações do Motorista",
                        "schema": {
                            "$ref": "#/definitions/drivers.DriverResponse"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/driver/update": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Atualiza os dados de um motorista.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Motoristas"
                ],
                "summary": "Atualizar um Motorista.",
                "parameters": [
                    {
                        "description": "Requisição de Motorista",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/drivers.UpdateDriverRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações do Motorista",
                        "schema": {
                            "$ref": "#/definitions/drivers.DriverResponse"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "drivers.CreateDriverRequest": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "cep": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "complement": {
                    "type": "string"
                },
                "cpf": {
                    "type": "string"
                },
                "license_category": {
                    "type": "string",
                    "enum": [
                        "a",
                        "b",
                        "c",
                        "d",
                        "e"
                    ]
                },
                "license_expiration_date": {
                    "type": "string"
                },
                "license_number": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "neighborhood": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "street_number": {
                    "type": "string"
                }
            }
        },
        "drivers.DriverResponse": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "cep": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "complement": {
                    "type": "string"
                },
                "cpf": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "license_category": {
                    "type": "string"
                },
                "license_expiration_date": {
                    "type": "string"
                },
                "license_number": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "neighborhood": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "street": {
                    "type": "string"
                },
                "street_number": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "drivers.UpdateDriverRequest": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "cep": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "complement": {
                    "type": "string"
                },
                "cpf": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "license_category": {
                    "type": "string",
                    "enum": [
                        "a",
                        "b",
                        "c",
                        "d",
                        "e"
                    ]
                },
                "license_expiration_date": {
                    "type": "string"
                },
                "license_number": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "neighborhood": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "street_number": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}